Placeholder textures, simple level generation, working health bar, smart virus
All of the viruses and cells are generated from prefabs in the
playerMovement script. The health bar now works properly. Viruses are
inteligent enough to decide between chasing the player, fleeing the
player, moving toward the nearest cell, and choosing a random direction.
Game near playability, need win/lose condition, minimap and the ability
to shoot.