Virus complete
The virus is in an Alpha worthy state. It behaves very well. It will not
leave the bounding box and bounces appropriately off of borders. The
virus typically heads in the rough direction of the nearest cell or
occasionally off in a random direction. When it contacts a cell it
attaches. It stays in the same position relative to the cell and matches
its velocity until the cell dies. There are comands to attack and flee
from the player as well. Attack will negatively affect the player once
shooting is developed to get them off.