Cells, death sounds, scoring, map, balance
Cleaned up everything. Cells no longer can escape the bounding box.
Viruses and cells make sounds on death. The score now is determined by
the time that has passed as well as the number of viruses. The minimap
has been resized and completed. The game is now balanced at around 40
seconds to a minute.  Object resizing to better suit the environment.