Bullet prefab fixed, lingering death particles, bullet explosion
The bullet prefab was remade. A simple rigidbody sphere with the
bulletmovement script. It's attached to the player but shoots wrong.
Viruses and Cells now create an appropriate color explosion on death
that lingers after they are destroyed. I've also created the bullet
explosion functionality. Currently you can right click to use it
centered on the player.  Once the bullet works it will occur where the
bullet is destroyed.